+++ Twist/src/com/secondhand/model/Obstacle.java
@@ -4,7 +4,7 @@ import com.badlogic.gdx.math.Vector2;
 
 public class Obstacle extends Entity {
 
-	public Obstacle(Vector2 position, int radius) {
+	public Obstacle(Vector2 position, float radius) {
 		super(position,radius);
 	}
